airflow/requirements.txt
astronomer-cosmos
dbt-trino
apache-airflow-providers-trino